requirements.txt
numpy
torch
torch-geometric
absl-py
